+++ api/src/controllers/batch/seal.ts
@@ -1,4 +1,4 @@
-// Copyright 2022 Touca, Inc. Subject to Apache-2.0 License.
+// Copyright 2023 Touca, Inc. Subject to Apache-2.0 License.
 
 import { NextFunction, Request, Response } from 'express'
 
@@ -11,21 +11,6 @@ import {
 } from '../../schemas/index.js'
 import { analytics, logger } from '../../utils/index.js'
 
-/**
- * @summary
- * Seal a given version of a suite in a team.
- *
- * @description
- * Seals a batch `batch` of a suite `suite` to prevent further submission
- * of results in the future.
- *
- * This function is designed to be called after the following middleware:
- *  - `isAuthenticated` to yield `user`
- *  - `hasTeam` to yield `team`
- *  - `isTeamMember`
- *  - `hasSuite` to yield `suite`
- *  - `hasBatch` to yield `batch`
- */
 export async function ctrlBatchSeal(
   req: Request,
   res: Response,
@@ -37,16 +22,12 @@ export async function ctrlBatchSeal(
   const batch = res.locals.batch as IBatchDocument
   const tuple = [team.slug, suite.slug, batch.slug].join('/')
   logger.debug('%s: %s: sealing', user.username, tuple)
-
-  // we are done if batch is already sealed
   if (batch.sealedAt) {
     logger.info('%s: %s: already sealed', user.username, tuple)
     return res.status(204).send()
   }
-
   await batchSeal(team, suite, batch)
   analytics.add_activity('batch:sealed', user)
   logger.info('%s: %s: sealed', user.username, tuple)
-
   return res.status(204).send()
 }