+++ sdk/cpp/src/client.cpp
@@ -175,7 +175,7 @@ void ClientImpl::seal() const {
         "client is not configured to contact the server");
   }
   const auto& response = _transport->post(
-      touca::detail::format("/batch/{}/{}/{}/seal2", _options.team,
+      touca::detail::format("/client/seal/{}/{}/{}", _options.team,
                             _options.suite, _options.version));
   if (response.status == -1) {
     throw touca::detail::runtime_error(