+++ sdk/cpp/src/runner.cpp
@@ -205,7 +205,8 @@ void Printer::print_progress(const unsigned index, const Status status,
 }
 
 void Printer::print_footer(const Statistics& stats, Timer& timer,
-                           const unsigned suiteSize) {
+                           const Workflow workflow,
+                           const RunnerOptions& options) {
   const auto duration = timer.count("__workflow__") / 1000.0;
   const auto report = [&](const Status state, const fmt::terminal_color color,
                           const std::string& name) {
@@ -218,8 +219,18 @@ void Printer::print_footer(const Statistics& stats, Timer& timer,
   report(Status::Pass, fmt::terminal_color::green, "passed");
   report(Status::Skip, fmt::terminal_color::yellow, "skipped");
   report(Status::Fail, fmt::terminal_color::red, "failed");
-  print("{} total\n", suiteSize);
+  print("{} total\n", static_cast<unsigned int>(workflow.testcases.size()));
   print("Time:       {:.2f} s\n", duration);
+  if (!options.web_url.empty()) {
+    print("Link:       {}/~/{}/{}/{}\n", options.web_url, options.team,
+          workflow.suite, workflow.version);
+  }
+  if (options.save_binary || options.save_json) {
+    const auto& results_dir =
+        touca::filesystem::path(options.output_directory) / workflow.suite /
+        workflow.version;
+    print("Results:    {}\n", results_dir.string());
+  }
 }
 
 void Runner::run_workflows() {
@@ -284,9 +295,7 @@ void Runner::run_workflow(const Workflow& workflow) {
     run_testcase(workflow, testcase, index++);
   }
   timer.toc("__workflow__");
-  printer.print_footer(stats, timer,
-                       static_cast<unsigned int>(workflow.testcases.size()));
-
+  printer.print_footer(stats, timer, workflow, options);
   if (!options.offline && stats.count(Status::Pass) != 0) {
     touca::seal();
   }