+++ sdk/java/touca/src/main/java/io/touca/core/Client.java
@@ -6,20 +6,15 @@ import com.google.flatbuffers.FlatBufferBuilder;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 import io.touca.TypeAdapter;
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.function.Consumer;
 
@@ -97,33 +92,6 @@ public class Client {
    */
   public String configurationError() {
     return this.configError;
-  }
-
-  /**
-   * Queries the Touca server for the list of testcases that are submitted to
-   * the baseline version of this suite.
-   *
-   * @return list of test cases of the baseline version of this suite
-   * @throws ToucaException when called on the client that is not configured to
-   *                        communicate with the Touca server.
-   */
-  public List<String> getTestcases() {
-    if (this.transport == null) {
-      throw new ToucaException("client not configured to perform this operation");
-    }
-    final Transport.Response response = transport.getRequest(
-        String.format("/client/element/%s/%s", options.team, options.suite));
-    if (response.code != HttpURLConnection.HTTP_OK) {
-      throw new ToucaException("failed to obtain list of test cases");
-    }
-    final JsonElement content = JsonParser.parseString(response.content);
-    final JsonArray array = content.getAsJsonArray();
-    final List<String> elements = new ArrayList<>();
-    for (int i = 0; i < array.size(); i++) {
-      final JsonObject element = array.get(i).getAsJsonObject();
-      elements.add(element.get("name").getAsString());
-    }
-    return elements;
   }
 
   /**
@@ -334,7 +302,7 @@ public class Client {
       throw new ToucaException("client is not configured to contact the server");
     }
     final Transport.Response response = this.transport.postRequest(
-        String.format("/batch/%s/%s/%s/seal2", options.team, options.suite, options.version),
+        String.format("/client/seal/%s/%s/%s", options.team, options.suite, options.version),
         "application/json", new byte[0]);
     if (response.code == HttpURLConnection.HTTP_FORBIDDEN) {
       throw new ToucaException("client is not authenticated");