+++ sdk/js/Changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## v1.6.4
+
+Features:
+
+- Test runner output should include link to results on the server (#581)
+
+Improvements:
+
+- Update dependencies (#581)
+
 ## v1.6.3
 
 Improvements: